Split generator class in two
