+++ decidim-admin/app/views/decidim/admin/components/_component.html.erb
@@ -32,9 +32,11 @@
       <% end %>
     <% end %>
 
-    <%= form_tag(decidim_admin.pseudomizations_path, method: :post, class: "action-icon action-icon--configure") do %>
-      <%= hidden_field_tag(:component_id, component.id) %>
-      <%= submit_tag(icon("key")) %>
+    <% if allowed_to? :pseudomize, :component, component: component %>
+      <%= form_tag(decidim_admin.pseudomizations_path, method: :post, class: "action-icon action-icon--configure") do %>
+        <%= hidden_field_tag(:component_id, component.id) %>
+        <%= submit_tag(icon("shield")) %>
+      <% end %>
     <% end %>
 
     <% if allowed_to? :destroy, :component, component: component %>
